legacy nuxt/api/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-ckeditor-5
django-taggit
django-taggit-serializer
pillow
requests
Faker
